requirements.txt
requests
reportlab
pillow
python-dotenv
customtkinter
